requirements_v2.txt
fastapi
uvicorn[standard]
celery[redis]
redis
python-dotenv
sqlalchemy
pymysql
pydantic-settings
loguru
openai

